fix(rrule): extract parseRRule into leaf module — unblocks Vercel/Turbopack build
PR #1430 Vercel build failed because the refactor that pulled parseRRule
into client-bundle import chains (format.ts re-exports collectKennelWeekdays
from schedule-season.ts, which imported parseRRule from
@/adapters/static-schedule/adapter, which transitively pulls in
@/adapters/utils → safe-fetch → ssrf-dns → node:dns/promises). Turbopack
correctly rejects `node:dns/promises` from a client bundle:

  > the chunking context (unknown) does not support external modules
  > (request: node:dns/promises)

Root cause: parseRRule is a pure function but lived in a module file with
server-only transitive deps.

Fix: extract parseRRule + its helpers (parseInterval / parseByDay /
parseByMonthDay / parseByMonth, SUPPORTED_FREQS, ParsedRRule type) into
a new leaf module src/adapters/static-schedule/rrule-parser.ts with NO
server-only imports. The static-schedule adapter re-exports parseRRule
from the leaf module so existing import paths still work (Travel Mode
projection, the adapter test surface). format.ts + schedule-season.ts
now import directly from the leaf module.

WEEKDAY_NAMES (a Record<string, number> for RFC 5545 day-token lookup)
was the only server-tainted import parseRRule pulled in; the leaf
module inlines a local RRULE_WEEKDAY_TO_NUM constant (7 entries, the
RFC 5545 abbreviation subset parseRRule actually consumes — the full
WEEKDAY_NAMES table also has full-word keys that the adapter shape
doesn't need).

Verification:
- npx tsc --noEmit clean
- npm test: 6505 tests pass (no behavior change, just refactor)
- Client-bundle import chain confirmed clean: KennelDirectory →
  format → schedule-season → rrule-parser has zero server-only paths.
