+++ src/adapters/static-schedule/adapter.ts
@@ -8,7 +8,13 @@
 
 import type { Source } from "@/generated/prisma/client";
 import type { SourceAdapter, RawEventData, ScrapeResult } from "../types";
-import { validateSourceConfig, WEEKDAY_NAMES } from "../utils";
+import { validateSourceConfig } from "../utils";
+import { parseRRule, type ParsedRRule } from "./rrule-parser";
+
+// Re-export the pure RRULE parser from its leaf module for callers that still
+// import from this file (Travel Mode projection, the existing test surface).
+export { parseRRule } from "./rrule-parser";
+export type { ParsedRRule };
 import {
   generateLunarOccurrences,
   ANCHOR_WEEKDAYS,
@@ -50,134 +56,6 @@ export interface StaticScheduleConfig {
    * as literal text on per-source rows whose RRULE matches that slot.
    */
   titleTemplate?: string;
-}
-
-/** Supported FREQ values. Other values (DAILY, YEARLY, etc.) are rejected. */
-const SUPPORTED_FREQS = new Set(["WEEKLY", "MONTHLY"]);
-
-/**
- * Parse INTERVAL from RRULE parts. Returns 1 if not specified.
- * @throws {Error} If INTERVAL is not a finite positive integer.
- */
-function parseInterval(parts: Record<string, string>): number {
-  if (!parts.INTERVAL) return 1;
-  const interval = Number.parseInt(parts.INTERVAL, 10);
-  if (!Number.isFinite(interval) || interval < 1) {
-    throw new Error(`Invalid INTERVAL: ${parts.INTERVAL} (must be >= 1)`);
-  }
-  return interval;
-}
-
-/**
- * Parse BYDAY from RRULE parts. Supports formats like "SA", "2SA" (2nd Saturday),
- * "-1FR" (last Friday). Returns undefined if BYDAY is not present.
- * @throws {Error} If BYDAY format is invalid, day abbreviation is unknown, or nth is 0.
- */
-function parseByDay(parts: Record<string, string>): { day: number; nth?: number } | undefined {
-  if (!parts.BYDAY) return undefined;
-  const match = /^(-?\d+)?([A-Z]{2})$/.exec(parts.BYDAY);
-  if (!match) throw new Error(`Invalid BYDAY: ${parts.BYDAY}`);
-  const dayNum = WEEKDAY_NAMES[match[2]];
-  if (dayNum === undefined) throw new Error(`Unknown day: ${match[2]}`);
-  if (match[1]) {
-    const nth = Number.parseInt(match[1], 10);
-    if (nth === 0) throw new Error("BYDAY nth position cannot be 0");
-    return { day: dayNum, nth };
-  }
-  return { day: dayNum };
-}
-
-/**
- * Parse BYMONTHDAY from RRULE parts. Returns undefined if not present.
- * @throws {Error} If BYMONTHDAY is not a valid day of month (1-31).
- */
-function parseByMonthDay(parts: Record<string, string>): number | undefined {
-  if (!parts.BYMONTHDAY) return undefined;
-  const byMonthDay = Number.parseInt(parts.BYMONTHDAY, 10);
-  if (!Number.isFinite(byMonthDay) || byMonthDay < 1 || byMonthDay > 31) {
-    throw new Error(`Invalid BYMONTHDAY: ${parts.BYMONTHDAY} (must be 1-31)`);
-  }
-  return byMonthDay;
-}
-
-/**
- * Parse BYMONTH from RRULE parts. Comma-separated list of month numbers (1-12)
- * per RFC 5545 §3.3.10. Returns a deduplicated, sorted list, or undefined if not present.
- * @throws {Error} If the list is empty or contains values outside 1-12.
- */
-function parseByMonth(parts: Record<string, string>): number[] | undefined {
-  if (!parts.BYMONTH) return undefined;
-  const tokens = parts.BYMONTH.split(",").map((t) => t.trim());
-  if (tokens.length === 0 || tokens.some((t) => t.length === 0)) {
-    throw new Error(`Invalid BYMONTH: ${parts.BYMONTH} (must be a comma-separated list of months 1-12)`);
-  }
-  const months = new Set<number>();
-  // RFC 5545 §3.3.10: monthnum = 1*2DIGIT — accept "5" and "05" but reject
-  // "005", "5.5", "5abc". Range check after parse handles "13", "99".
-  for (const token of tokens) {
-    if (!/^\d{1,2}$/.test(token)) {
-      throw new Error(`Invalid BYMONTH: ${parts.BYMONTH} (each value must be an integer 1-12)`);
-    }
-    const month = Number.parseInt(token, 10);
-    if (month < 1 || month > 12) {
-      throw new Error(`Invalid BYMONTH: ${parts.BYMONTH} (each value must be an integer 1-12)`);
-    }
-    months.add(month);
-  }
-  return [...months].sort((a, b) => a - b);
-}
-
-/**
- * Parse an RRULE string into structured parts.
- * Supports: FREQ (WEEKLY|MONTHLY), BYDAY (with optional nth prefix), INTERVAL, BYMONTHDAY,
- * BYMONTH (comma-separated list of months 1-12, RFC 5545 §3.3.10).
- * Whitespace around semicolons and equals signs is trimmed.
- *
- * @throws {Error} On missing FREQ, unsupported FREQ, invalid INTERVAL/BYMONTHDAY/BYMONTH/BYDAY values.
- */
-export function parseRRule(rrule: string): {
-  freq: string;
-  interval: number;
-  byDay?: { day: number; nth?: number };
-  byMonthDay?: number;
-  byMonth?: number[];
-} {
-  const parts: Record<string, string> = {};
-  for (const segment of rrule.split(";")) {
-    const eqIdx = segment.indexOf("=");
-    if (eqIdx < 0) continue;
-    const key = segment.slice(0, eqIdx).trim().toUpperCase();
-    const value = segment.slice(eqIdx + 1).trim().toUpperCase();
-    if (key && value) parts[key] = value;
-  }
-
-  if (!parts.FREQ) throw new Error("RRULE missing FREQ");
-  const freq = parts.FREQ;
-  if (!SUPPORTED_FREQS.has(freq)) {
-    throw new Error(`Unsupported FREQ: ${freq} (supported: WEEKLY, MONTHLY)`);
-  }
-
-  // #1390: BYSETPOS is RFC 5545 valid but THIS parser doesn't honor it — and
-  // silently ignoring it caused Hebe H3 to ship "3rd Saturday" RRULEs that
-  // generated 1st-Saturday events instead. Fail loud so admins can't introduce
-  // the same drift again. The fix is BYDAY nth-prefix (e.g. BYDAY=1SA).
-  if (parts.BYSETPOS !== undefined) {
-    throw new Error(
-      "BYSETPOS is not supported (parser silently ignores it). " +
-        'Use BYDAY nth-prefix instead: BYDAY=1SA for "1st Saturday", BYDAY=-1FR for "last Friday".',
-    );
-  }
-
-  const interval = parseInterval(parts);
-  const byDay = parseByDay(parts);
-  const byMonthDay = parseByMonthDay(parts);
-  const byMonth = parseByMonth(parts);
-
-  if (freq === "WEEKLY" && !byDay) {
-    throw new Error("WEEKLY RRULE requires BYDAY");
-  }
-
-  return { freq, interval, byDay, byMonthDay, byMonth };
 }
 
 /**
